+++ server/src/internal/customers/attach/attachUtils/handleAttachErrors.ts
@@ -3,9 +3,7 @@ import {
 	AttachBranch,
 	type AttachConfig,
 	BillingType,
-	cusProductToProcessorType,
 	ErrCode,
-	ProcessorType,
 	RecaseError,
 	TierBehavior,
 	type UsagePriceConfig,
@@ -16,11 +14,11 @@ import {
 	getEntOptions,
 	getPriceEntitlement,
 	priceIsOneOffAndTiered,
-	pricesOnlyOneOff,
 } from "@/internal/products/prices/priceUtils.js";
 import { notNullish, nullOrUndefined } from "@/utils/genUtils.js";
 import type { AttachParams } from "../../cusProducts/AttachParams.js";
 import type { AttachFlags } from "../models/AttachFlags.js";
+import { handleExternalPSPErrors } from "./handleAttachErrors/handleExternalPSPErrors.js";
 import { handleMultiAttachErrors } from "./handleAttachErrors/handleMultiAttachErrors.js";
 
 const handleNonCheckoutErrors = ({
@@ -151,40 +149,6 @@ export const handleCustomPaymentMethodErrors = ({
 				"This customer is billed outside of Stripe, please use the origin platform to manage their billing.",
 		});
 	} else if (attachParams.customer.processors?.vercel?.installation_id) {
-		throw new RecaseError({
-			message:
-				"This customer is billed outside of Stripe, please use the origin platform to manage their billing.",
-		});
-	}
-};
-
-export const handleExternalPSPErrors = ({
-	attachParams,
-	strict = false,
-}: {
-	attachParams: AttachParams;
-	/**
-	 * When true, never bypass the cross-processor guard. Use for MultiAttach
-	 * where the customer's whole subscription state could change.
-	 */
-	strict?: boolean;
-}) => {
-	// Safe path: a single-product attach for a true one-off product can mix
-	// across processors. One-offs create a parallel cus_product and never
-	// replace an existing subscription, so a customer with an active RC sub
-	// can still buy a Stripe-billed top-up (and vice versa).
-	const oneOffEscape =
-		!strict &&
-		attachParams.products.length === 1 &&
-		pricesOnlyOneOff(attachParams.prices);
-
-	if (oneOffEscape) return;
-
-	if (
-		attachParams.customer.customer_products.some(
-			(cp) => cusProductToProcessorType(cp) !== ProcessorType.Stripe,
-		)
-	) {
 		throw new RecaseError({
 			message:
 				"This customer is billed outside of Stripe, please use the origin platform to manage their billing.",
