+++ server/src/internal/customers/attach/attachUtils/handleAttachErrors/handleCheckoutErrors.ts
@@ -1,9 +1,7 @@
 import type { AttachBranch } from "@autumn/shared";
 import type { AttachParams } from "@/internal/customers/cusProducts/AttachParams.js";
-import {
-	handleCustomPaymentMethodErrors,
-	handleExternalPSPErrors,
-} from "../handleAttachErrors.js";
+import { handleCustomPaymentMethodErrors } from "../handleAttachErrors.js";
+import { handleExternalPSPErrors } from "./handleExternalPSPErrors.js";
 
 export const handleCheckoutErrors = ({
 	attachParams,
